+++ optim/src/main.py
@@ -1,7 +1,7 @@
 import uuid
 from fastapi import FastAPI
 from typing import Dict, List, Union
-from app.optimize_round_with_focus_score_prediction import optimize_round_from_dynamodb_data_with_focus_score_prediction
+from app.optimize_round_from_dynamodb_data_with_focus_score_prediction import optimize_round_from_dynamodb_data_with_focus_score_prediction
 from app.optimize_round_from_requestbody_data import optimize_round_from_requestbody_data
 from app.optimize_session_from_dynamodb_data import optimize_session_from_dynamodb_data
 from app.optimize_session_from_requestbody_data import optimize_session_from_requestbody_data
